orchestrator/stt/requirements.txt
fastapi
python-multipart
uvicorn
openai-whisper 